+++ src/mongo/util/dns_query_test.cpp
@@ -110,7 +110,8 @@ TEST(MongoDnsQuery, basic) {
 
     // As long as enough tests pass, we're okay -- this means that a single DNS name server drift
     // won't cause a BF -- when enough fail, then we can rebuild the list in one pass.
-    const std::size_t kPassingRate = sizeof(tests) / sizeof(tests[0]) * kPassingPercentage;
+    const std::size_t kPassingRate =
+        static_cast<std::size_t>(sizeof(tests) / sizeof(tests[0])) * kPassingPercentage;
     ASSERT_GTE(resolution_count, kPassingRate);
 }
